Code Refactoring.
Refactored the code with a better interface and better
organization of files. It now follows the standard interface
based coding standard followed by other C libraries, i.e.,
interface in aes.h and implementation in aes.c. Earlier all the
code was in aes.h. It was stupidly done by me as I didn't
understand the concept of C interface.